Workout log works fine and Gym functionality

+++ src/WorkoutLogController.java
@@ -10,6 +10,18 @@ public class WorkoutLogController {
     private Statement stmt;
     private ResultSet rs;
     private ConnectionManager conMag = new ConnectionManager();
+    public Profile profile;
+
+    public WorkoutLogController(Profile p) {
+
+        con = conMag.getConnection();
+        try {
+            stmt = con.createStatement();
+            this.profile = p;
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
+        }
+    }
 
     public WorkoutLogController() {
         con = conMag.getConnection();
@@ -36,8 +48,17 @@ public class WorkoutLogController {
     public ArrayList<Exercise> getLogExercises(){
         ArrayList<Exercise> logExercises = new ArrayList<>();
         try {
+
             //TODO: Create list of exercises and use GUI to populate list.
-            rs = stmt.executeQuery("SELECT * FROM tbl_workoutLog");
+            String queryPrev = "SELECT user_id FROM users WHERE user_name =";
+            queryPrev += "'"+this.profile.getUserName()+"';";
+            ResultSet rs = stmt.executeQuery(queryPrev);
+            rs.next();
+
+
+            int id_rs = rs.getInt("user_id");
+
+            rs = stmt.executeQuery("SELECT * FROM tbl_workoutLog WHERE fk_user_id = '"+id_rs+"' ");
             while(rs.next()){
                 logExercises.add(new Exercise(rs.getString("exerciseName"), rs.getString("description"), rs.getDouble("weight"), rs.getInt("sets"), rs.getInt("reps")));
             }
@@ -49,7 +70,15 @@ public class WorkoutLogController {
 
     public boolean addExerciseToLog(Exercise e){
         try {
-            String query = "INSERT INTO tbl_workoutLog (`exerciseName`, `description`, `weight`, `sets`, `reps`) VALUES ('"+e.getName()+"', '"+e.getDescription()+"', '"+e.getWeight()+"', '"+e.getSets()+"', '"+e.getReps()+"');";
+            String queryPrev = "SELECT user_id FROM users WHERE user_name =";
+            queryPrev += "'"+this.profile.getUserName()+"';";
+            ResultSet rs = stmt.executeQuery(queryPrev);
+            rs.next();
+
+
+            int id_rs = rs.getInt("user_id");
+
+            String query = "INSERT INTO tbl_workoutLog (`exerciseName`, `description`, `weight`, `sets`, `reps`, `fk_user_id`) VALUES ('"+e.getName()+"', '"+e.getDescription()+"', '"+e.getWeight()+"', '"+e.getSets()+"', '"+e.getReps()+"', '"+id_rs+"');";
             stmt.executeUpdate(query);
             return true;
         } catch (SQLException throwables) {
@@ -60,7 +89,14 @@ public class WorkoutLogController {
 
     public boolean removeExerciseFromLog(Exercise e){
         try {
-            String query = "DELETE FROM tbl_workoutLog WHERE exerciseName = '" + e.getName() + "'AND weight = '" + e.getWeight() + "'";
+            String queryPrev = "SELECT user_id FROM users WHERE user_name =";
+            queryPrev += "'"+this.profile.getUserName()+"';";
+            ResultSet rs = stmt.executeQuery(queryPrev);
+            rs.next();
+
+
+            int id_rs = rs.getInt("user_id");
+            String query = "DELETE FROM tbl_workoutLog WHERE exerciseName = '" + e.getName() + "'AND weight = '" + e.getWeight() + "' AND fk_user_id = '" + id_rs + "'";
             stmt.executeUpdate(query);
             return true;
         } catch (SQLException throwables) {
@@ -70,8 +106,16 @@ public class WorkoutLogController {
     }
 
     public boolean updateExercise(Exercise e){
+
         try {
-            String query = "UPDATE tbl_workoutLog SET reps = '"+e.getReps()+"', sets = '"+e.getSets()+"', weight = '"+e.getWeight()+"' WHERE exerciseName = '"+e.getName()+"'";
+            String queryPrev = "SELECT user_id FROM users WHERE user_name =";
+            queryPrev += "'"+this.profile.getUserName()+"';";
+            ResultSet rs = stmt.executeQuery(queryPrev);
+            rs.next();
+
+
+            int id_rs = rs.getInt("user_id");
+            String query = "UPDATE tbl_workoutLog SET reps = '"+e.getReps()+"', sets = '"+e.getSets()+"', weight = '"+e.getWeight()+"' WHERE exerciseName = '"+e.getName()+"' AND fk_user_id = '"+ id_rs +"'";
             stmt.executeUpdate(query);
             return true;
         } catch (SQLException throwables) {
@@ -82,7 +126,15 @@ public class WorkoutLogController {
 
     public boolean resetLog(){
         try {
-            String query = "DELETE FROM tbl_workoutLog";
+
+            String queryPrev = "SELECT user_id FROM users WHERE user_name =";
+            queryPrev += "'"+this.profile.getUserName()+"';";
+            ResultSet rs = stmt.executeQuery(queryPrev);
+            rs.next();
+
+
+            int id_rs = rs.getInt("user_id");
+            String query = "DELETE FROM tbl_workoutLog WHERE fk_user_id = '"+id_rs+"'";
             stmt.executeUpdate(query);
             return true;
         } catch (SQLException throwables) {
